moved all the test to common
