+++ document-store/src/integrationTest/java/org/hypertrace/core/documentstore/DocStoreTest.java
@@ -3,30 +3,15 @@ package org.hypertrace.core.documentstore;
 import static org.hypertrace.core.documentstore.BulkArrayValueUpdateRequest.Operation.ADD;
 import static org.hypertrace.core.documentstore.BulkArrayValueUpdateRequest.Operation.REMOVE;
 import static org.hypertrace.core.documentstore.BulkArrayValueUpdateRequest.Operation.SET;
-import static org.hypertrace.core.documentstore.expression.operators.AggregationOperator.AVG;
-import static org.hypertrace.core.documentstore.expression.operators.AggregationOperator.COUNT;
-import static org.hypertrace.core.documentstore.expression.operators.AggregationOperator.DISTINCT_COUNT;
-import static org.hypertrace.core.documentstore.expression.operators.AggregationOperator.MAX;
-import static org.hypertrace.core.documentstore.expression.operators.AggregationOperator.MIN;
-import static org.hypertrace.core.documentstore.expression.operators.AggregationOperator.SUM;
-import static org.hypertrace.core.documentstore.expression.operators.FunctionOperator.MULTIPLY;
-import static org.hypertrace.core.documentstore.expression.operators.LogicalOperator.AND;
-import static org.hypertrace.core.documentstore.expression.operators.LogicalOperator.OR;
-import static org.hypertrace.core.documentstore.expression.operators.RelationalOperator.EQ;
-import static org.hypertrace.core.documentstore.expression.operators.RelationalOperator.GT;
-import static org.hypertrace.core.documentstore.expression.operators.RelationalOperator.GTE;
-import static org.hypertrace.core.documentstore.expression.operators.RelationalOperator.IN;
-import static org.hypertrace.core.documentstore.expression.operators.RelationalOperator.LTE;
-import static org.hypertrace.core.documentstore.expression.operators.RelationalOperator.NEQ;
-import static org.hypertrace.core.documentstore.expression.operators.SortOrder.ASC;
-import static org.hypertrace.core.documentstore.expression.operators.SortOrder.DESC;
 import static org.hypertrace.core.documentstore.utils.CreateUpdateTestThread.FAILURE;
 import static org.hypertrace.core.documentstore.utils.CreateUpdateTestThread.SUCCESS;
-import static org.hypertrace.core.documentstore.utils.Utils.assertDocsAndSizeEqual;
-import static org.hypertrace.core.documentstore.utils.Utils.convertDocumentToMap;
-import static org.hypertrace.core.documentstore.utils.Utils.convertJsonToMap;
-import static org.hypertrace.core.documentstore.utils.Utils.createDocumentsFromResource;
-import static org.hypertrace.core.documentstore.utils.Utils.readFileFromResource;
+import static org.hypertrace.core.documentstore.utils.Utils.MONGO_CREATED_TIME_KEY;
+import static org.hypertrace.core.documentstore.utils.Utils.MONGO_LAST_UPDATED_TIME_KEY;
+import static org.hypertrace.core.documentstore.utils.Utils.MONGO_LAST_UPDATE_TIME_KEY;
+import static org.hypertrace.core.documentstore.utils.Utils.MONGO_STORE;
+import static org.hypertrace.core.documentstore.utils.Utils.POSTGRES_CREATED_AT;
+import static org.hypertrace.core.documentstore.utils.Utils.POSTGRES_STORE;
+import static org.hypertrace.core.documentstore.utils.Utils.POSTGRES_UPDATED_AT;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -55,20 +40,8 @@ import org.apache.commons.lang3.RandomUtils;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.bson.codecs.configuration.CodecConfigurationException;
 import org.hypertrace.core.documentstore.Filter.Op;
-import org.hypertrace.core.documentstore.expression.impl.AggregateExpression;
-import org.hypertrace.core.documentstore.expression.impl.ConstantExpression;
-import org.hypertrace.core.documentstore.expression.impl.FunctionExpression;
-import org.hypertrace.core.documentstore.expression.impl.IdentifierExpression;
-import org.hypertrace.core.documentstore.expression.impl.LogicalExpression;
-import org.hypertrace.core.documentstore.expression.impl.RelationalExpression;
-import org.hypertrace.core.documentstore.expression.impl.UnnestExpression;
 import org.hypertrace.core.documentstore.mongo.MongoDatastore;
 import org.hypertrace.core.documentstore.postgres.PostgresDatastore;
-import org.hypertrace.core.documentstore.query.Pagination;
-import org.hypertrace.core.documentstore.query.Selection;
-import org.hypertrace.core.documentstore.query.SelectionSpec;
-import org.hypertrace.core.documentstore.query.Sort;
-import org.hypertrace.core.documentstore.query.SortingSpec;
 import org.hypertrace.core.documentstore.utils.CreateUpdateTestThread;
 import org.hypertrace.core.documentstore.utils.CreateUpdateTestThread.Operation;
 import org.hypertrace.core.documentstore.utils.Utils;
@@ -85,25 +58,8 @@ import org.testcontainers.shaded.com.google.common.collect.Maps;
 import org.testcontainers.utility.DockerImageName;
 
 public class DocStoreTest {
-
-  public static final String MONGO_STORE = "Mongo";
-  public static final String POSTGRES_STORE = "Postgres";
   private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
   private static final String COLLECTION_NAME = "myTest";
-
-  /*
-   * These 3 fields should be automatically created when upserting a doc.
-   * There are downstream services that depends on this. The test should verify that
-   * the string is not changed.
-   */
-  private static final String MONGO_LAST_UPDATE_TIME_KEY = "_lastUpdateTime";
-  private static final String MONGO_LAST_UPDATED_TIME_KEY = "lastUpdatedTime";
-  private static final String MONGO_CREATED_TIME_KEY = "createdTime";
-  /** Postgres related time fields */
-  public static final String POSTGRES_UPDATED_AT = "updated_at";
-
-  public static final String POSTGRES_CREATED_AT = "created_at";
-
   private static Map<String, Datastore> datastoreMap;
 
   private static GenericContainer<?> mongo;
@@ -2359,582 +2315,6 @@ public class DocStoreTest {
 
   @ParameterizedTest
   @MethodSource("databaseContextProvider")
-  public void testQueryV1ForSimpleWhereClause(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    // query docs
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .setFilter(
-                RelationalExpression.of(
-                    IdentifierExpression.of("quantity"), NEQ, ConstantExpression.of(10)))
-            .build();
-
-    Iterator<Document> iterator = collection.aggregate(query);
-    assertSizeAndDocsEqual(dataStoreName, iterator, 6, "mongo/simple_filter_quantity_neq_10.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testQueryV1FilterWithNestedFiled(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    // query docs
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .setFilter(
-                LogicalExpression.builder()
-                    .operator(AND)
-                    .operand(
-                        RelationalExpression.of(
-                            IdentifierExpression.of("quantity"), GT, ConstantExpression.of(5)))
-                    .operand(
-                        RelationalExpression.of(
-                            IdentifierExpression.of("props.seller.address.city"),
-                            EQ,
-                            ConstantExpression.of("Kolkata")))
-                    .build())
-            .build();
-
-    Iterator<Document> iterator = collection.aggregate(query);
-    assertSizeAndDocsEqual(
-        dataStoreName, iterator, 1, "mongo/test_nest_field_filter_response.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testQueryV1ForFilterWithLogicalExpressionAndOr(String dataStoreName)
-      throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    // query docs
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .setFilter(
-                LogicalExpression.builder()
-                    .operand(
-                        RelationalExpression.of(
-                            IdentifierExpression.of("price"), EQ, ConstantExpression.of(10)))
-                    .operator(OR)
-                    .operand(
-                        LogicalExpression.builder()
-                            .operand(
-                                RelationalExpression.of(
-                                    IdentifierExpression.of("quantity"),
-                                    GTE,
-                                    ConstantExpression.of(5)))
-                            .operator(AND)
-                            .operand(
-                                RelationalExpression.of(
-                                    IdentifierExpression.of("quantity"),
-                                    LTE,
-                                    ConstantExpression.of(10)))
-                            .build())
-                    .build())
-            .build();
-
-    Iterator<Document> resultDocs = collection.aggregate(query);
-    assertSizeAndDocsEqual(
-        dataStoreName, resultDocs, 6, "mongo/filter_with_logical_and_or_operator.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testQueryV1ForSelectionExpression(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    // query docs
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .setFilter(
-                RelationalExpression.of(
-                    IdentifierExpression.of("price"), EQ, ConstantExpression.of(10)))
-            .addSelection(IdentifierExpression.of("item"))
-            .addSelection(IdentifierExpression.of("props.brand"))
-            .addSelection(IdentifierExpression.of("props.seller.name"))
-            .addSelection(
-                FunctionExpression.builder()
-                    .operand(IdentifierExpression.of("price"))
-                    .operator(MULTIPLY)
-                    .operand(IdentifierExpression.of("quantity"))
-                    .build(),
-                "total")
-            .build();
-
-    Iterator<Document> resultDocs = collection.aggregate(query);
-    assertSizeAndDocsEqual(
-        dataStoreName, resultDocs, 2, "mongo/test_selection_expression_result.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testQueryV1FunctionalSelectionExpressionWithNestedFieldWithAlias(String dataStoreName)
-      throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    // query docs
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .setFilter(
-                RelationalExpression.of(
-                    IdentifierExpression.of("price"), EQ, ConstantExpression.of(10)))
-            .addSelection(IdentifierExpression.of("item"))
-            .addSelection(IdentifierExpression.of("props.brand"), "props_brand")
-            .addSelection(IdentifierExpression.of("props.seller.name"), "props_seller_name")
-            .addSelection(
-                FunctionExpression.builder()
-                    .operand(IdentifierExpression.of("price"))
-                    .operator(MULTIPLY)
-                    .operand(IdentifierExpression.of("quantity"))
-                    .build(),
-                "total")
-            .build();
-
-    Iterator<Document> resultDocs = collection.aggregate(query);
-    assertSizeAndDocsEqual(
-        dataStoreName,
-        resultDocs,
-        2,
-        "mongo/test_selection_expression_nested_fields_alias_result.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testQueryV1AggregationExpression(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .setFilter(
-                RelationalExpression.of(
-                    IdentifierExpression.of("price"), GT, ConstantExpression.of(5)))
-            .addSelection(IdentifierExpression.of("item"))
-            .addSelection(
-                AggregateExpression.of(AVG, IdentifierExpression.of("quantity")), "qty_avg")
-            .addSelection(
-                AggregateExpression.of(COUNT, IdentifierExpression.of("quantity")), "qty_count")
-            .addSelection(
-                AggregateExpression.of(DISTINCT_COUNT, IdentifierExpression.of("quantity")),
-                "qty_distinct_count")
-            .addSelection(
-                AggregateExpression.of(SUM, IdentifierExpression.of("quantity")), "qty_sum")
-            .addSelection(
-                AggregateExpression.of(MIN, IdentifierExpression.of("quantity")), "qty_min")
-            .addSelection(
-                AggregateExpression.of(MAX, IdentifierExpression.of("quantity")), "qty_max")
-            .addAggregation(IdentifierExpression.of("item"))
-            .build();
-
-    Iterator<Document> resultDocs = collection.aggregate(query);
-    assertSizeAndDocsEqual(
-        dataStoreName, resultDocs, 3, "mongo/test_aggregation_expression_result.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testQueryV1AggregationFilter(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .addSelection(
-                AggregateExpression.of(DISTINCT_COUNT, IdentifierExpression.of("quantity")),
-                "qty_count")
-            .addSelection(IdentifierExpression.of("item"))
-            .addAggregation(IdentifierExpression.of("item"))
-            .setAggregationFilter(
-                RelationalExpression.of(
-                    IdentifierExpression.of("qty_count"), LTE, ConstantExpression.of(10)))
-            .build();
-
-    Iterator<Document> resultDocs = collection.aggregate(query);
-    assertSizeAndDocsEqual(dataStoreName, resultDocs, 4, "mongo/distinct_count_response.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testQueryV1AggregationFilterWithWhereClause(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .addSelection(
-                AggregateExpression.of(DISTINCT_COUNT, IdentifierExpression.of("quantity")),
-                "qty_count")
-            .addSelection(IdentifierExpression.of("item"))
-            .addAggregation(IdentifierExpression.of("item"))
-            .setFilter(
-                RelationalExpression.of(
-                    IdentifierExpression.of("price"), LTE, ConstantExpression.of(7.5)))
-            .setAggregationFilter(
-                RelationalExpression.of(
-                    IdentifierExpression.of("qty_count"), LTE, ConstantExpression.of(10)))
-            .build();
-
-    Iterator<Document> resultDocs = collection.aggregate(query);
-    assertSizeAndDocsEqual(
-        dataStoreName, resultDocs, 2, "mongo/test_aggr_filter_and_where_filter_result.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testUnnestWithoutPreserveNullAndEmptyArrays(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .addSelection(IdentifierExpression.of("item"))
-            .addSelection(IdentifierExpression.of("price"))
-            .addSelection(IdentifierExpression.of("sales.city"))
-            .addSelection(IdentifierExpression.of("sales.medium.type"))
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales"), false))
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales.medium"), false))
-            .build();
-
-    Iterator<Document> resultDocs = collection.aggregate(query);
-    assertSizeAndDocsEqual(
-        dataStoreName, resultDocs, 11, "mongo/unwind_not_preserving_selection_response.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testUnnestWithoutPreserveNullAndEmptyArraysWithFilters(String dataStoreName)
-      throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .addSelection(IdentifierExpression.of("item"))
-            .addSelection(IdentifierExpression.of("sales.city"))
-            .addSelection(IdentifierExpression.of("sales.medium.type"))
-            .setFilter(
-                RelationalExpression.of(
-                    IdentifierExpression.of("quantity"), NEQ, ConstantExpression.of(10)))
-            .addFromClause(
-                UnnestExpression.builder()
-                    .identifierExpression(IdentifierExpression.of("sales"))
-                    .preserveNullAndEmptyArrays(false)
-                    .filterTypeExpression(
-                        RelationalExpression.of(
-                            IdentifierExpression.of("sales.city"),
-                            EQ,
-                            ConstantExpression.of("delhi")))
-                    .build())
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales.medium"), false))
-            .build();
-
-    Iterator<Document> resultDocs = collection.aggregate(query);
-    assertSizeAndDocsEqual(
-        dataStoreName, resultDocs, 3, "mongo/unwind_not_preserving_filter_response.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testUnnestWithPreserveNullAndEmptyArrays(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .addSelection(IdentifierExpression.of("item"))
-            .addSelection(IdentifierExpression.of("price"))
-            .addSelection(IdentifierExpression.of("sales.city"))
-            .addSelection(IdentifierExpression.of("sales.medium.type"))
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales"), true))
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales.medium"), true))
-            .build();
-
-    Iterator<Document> resultDocs = collection.aggregate(query);
-    assertSizeAndDocsEqual(
-        dataStoreName, resultDocs, 17, "mongo/unwind_preserving_selection_response.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testUnnestAndAggregate(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .addSelection(IdentifierExpression.of("sales.medium.type"))
-            .addAggregation(IdentifierExpression.of("sales.medium.type"))
-            .addSelection(
-                AggregateExpression.of(SUM, IdentifierExpression.of("sales.medium.volume")),
-                "totalSales")
-            // we don't want to consider entries where sales data is missing
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales"), false))
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales.medium"), false))
-            .addSort(IdentifierExpression.of("totalSales"), DESC)
-            .build();
-
-    Iterator<Document> iterator = collection.aggregate(query);
-    assertSizeAndDocsEqual(
-        dataStoreName, iterator, 3, "mongo/aggregate_on_nested_array_reponse.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testUnnestAndAggregate_preserveEmptyTrue(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    // include all documents in the result irrespective of `sales` field
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .addSelection(AggregateExpression.of(COUNT, IdentifierExpression.of("item")), "count")
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales"), true))
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales.medium"), true))
-            .build();
-
-    Iterator<Document> iterator = collection.aggregate(query);
-    assertSizeAndDocsEqual(
-        dataStoreName, iterator, 1, "mongo/unwind_preserving_empty_array_response.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testUnnest(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .addSelection(IdentifierExpression.of("item"))
-            .addSelection(IdentifierExpression.of("sales.city"))
-            .addSelection(IdentifierExpression.of("sales.medium"))
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales"), true))
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales.medium"), true))
-            .addSort(IdentifierExpression.of("item"), DESC)
-            .addSort(IdentifierExpression.of("sales.city"), DESC)
-            .addSort(IdentifierExpression.of("sales.medium.volume"), DESC)
-            .addSort(IdentifierExpression.of("sales.medium.type"), DESC)
-            .build();
-
-    Iterator<Document> iterator = collection.aggregate(query);
-    assertSizeAndDocsEqual(dataStoreName, iterator, 17, "mongo/unwind_response.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testUnnestAndAggregate_preserveEmptyFalse(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    // consider only those documents where sales field is missing
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .addSelection(AggregateExpression.of(COUNT, IdentifierExpression.of("item")), "count")
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales"), false))
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales.medium"), true))
-            .build();
-
-    Iterator<Document> iterator = collection.aggregate(query);
-    assertSizeAndDocsEqual(
-        dataStoreName, iterator, 1, "mongo/unwind_not_preserving_empty_array_response.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testFilterAndUnnest(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    RelationalExpression relationalExpression =
-        RelationalExpression.of(
-            IdentifierExpression.of("sales.city"), EQ, ConstantExpression.of("delhi"));
-
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .addSelection(IdentifierExpression.of("item"))
-            .addSelection(IdentifierExpression.of("sales.city"))
-            .addSelection(IdentifierExpression.of("sales.medium"))
-            .addFromClause(
-                UnnestExpression.builder()
-                    .identifierExpression(IdentifierExpression.of("sales"))
-                    .preserveNullAndEmptyArrays(true)
-                    .filterTypeExpression(relationalExpression)
-                    .build())
-            .addFromClause(UnnestExpression.of(IdentifierExpression.of("sales.medium"), true))
-            .addSort(IdentifierExpression.of("item"), DESC)
-            .addSort(IdentifierExpression.of("sales.city"), DESC)
-            .addSort(IdentifierExpression.of("sales.medium.volume"), DESC)
-            .addSort(IdentifierExpression.of("sales.medium.type"), DESC)
-            .build();
-
-    Iterator<Document> iterator = collection.aggregate(query);
-    assertSizeAndDocsEqual(dataStoreName, iterator, 7, "mongo/unwind_filter_response.json");
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testQueryV1DistinctCountWithSortingSpecs(String dataStoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(dataStoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .addSelection(
-                AggregateExpression.of(DISTINCT_COUNT, IdentifierExpression.of("quantity")),
-                "qty_count")
-            .addSelection(IdentifierExpression.of("item"))
-            .addAggregation(IdentifierExpression.of("item"))
-            .setAggregationFilter(
-                RelationalExpression.of(
-                    IdentifierExpression.of("qty_count"), LTE, ConstantExpression.of(1000)))
-            .addSort(IdentifierExpression.of("qty_count"), DESC)
-            .addSort(IdentifierExpression.of("item"), DESC)
-            .build();
-
-    Iterator<Document> resultDocs = collection.aggregate(query);
-    assertDocsAndSizeEqual(resultDocs, "mongo/distinct_count_response.json", 4);
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
-  public void testFindWithSortingAndPagination(String datastoreName) throws IOException {
-    Map<Key, Document> documents = createDocumentsFromResource("mongo/collection_data.json");
-    Datastore datastore = datastoreMap.get(datastoreName);
-    Collection collection = datastore.getCollection(COLLECTION_NAME);
-
-    // add docs
-    boolean result = collection.bulkUpsert(documents);
-    Assertions.assertTrue(result);
-    List<SelectionSpec> selectionSpecs =
-        List.of(
-            SelectionSpec.of(IdentifierExpression.of("item")),
-            SelectionSpec.of(IdentifierExpression.of("price")),
-            SelectionSpec.of(IdentifierExpression.of("quantity")),
-            SelectionSpec.of(IdentifierExpression.of("date")));
-    Selection selection = Selection.builder().selectionSpecs(selectionSpecs).build();
-
-    org.hypertrace.core.documentstore.query.Filter filter =
-        org.hypertrace.core.documentstore.query.Filter.builder()
-            .expression(
-                RelationalExpression.of(
-                    IdentifierExpression.of("item"),
-                    IN,
-                    ConstantExpression.ofStrings(List.of("Mirror", "Comb", "Shampoo", "Bottle"))))
-            .build();
-
-    Sort sort =
-        Sort.builder()
-            .sortingSpec(SortingSpec.of(IdentifierExpression.of("quantity"), DESC))
-            .sortingSpec(SortingSpec.of(IdentifierExpression.of("item"), ASC))
-            .build();
-
-    Pagination pagination = Pagination.builder().offset(1).limit(3).build();
-
-    org.hypertrace.core.documentstore.query.Query query =
-        org.hypertrace.core.documentstore.query.Query.builder()
-            .setSelection(selection)
-            .setFilter(filter)
-            .setSort(sort)
-            .setPagination(pagination)
-            .build();
-
-    Iterator<Document> resultDocs = collection.find(query);
-    Utils.assertDocsAndSizeEqual(
-        resultDocs, "mongo/filter_with_sorting_and_pagination_response.json", 3);
-  }
-
-  @ParameterizedTest
-  @MethodSource("databaseContextProvider")
   public void whenBulkUpdatingNonExistentRecords_thenExpectNothingToBeUpdatedOrCreated(
       String datastoreName) throws Exception {
     Datastore datastore = datastoreMap.get(datastoreName);
@@ -3098,28 +2478,6 @@ public class DocStoreTest {
     }
   }
 
-  private static void assertSizeAndDocsEqual(
-      String dataStoreName, Iterator<Document> documents, int expectedSize, String filePath)
-      throws IOException {
-    String fileContent = readFileFromResource(filePath).orElseThrow();
-    List<Map<String, Object>> expectedDocs = convertJsonToMap(fileContent);
-
-    List<Map<String, Object>> actualDocs = new ArrayList<>();
-    int actualSize = 0;
-    while (documents.hasNext()) {
-      Map<String, Object> doc = convertDocumentToMap(documents.next());
-      removesDateRelatedFields(dataStoreName, doc);
-      actualDocs.add(doc);
-      actualSize++;
-    }
-
-    long count =
-        expectedDocs.stream().filter(expectedDoc -> actualDocs.contains(expectedDoc)).count();
-
-    assertEquals(expectedSize, actualSize);
-    assertEquals(expectedSize, count);
-  }
-
   private static void assertSizeEqual(Iterator<Document> documents, int expectedSize) {
     int actualSize = 0;
     while (documents.hasNext()) {
@@ -3127,16 +2485,5 @@ public class DocStoreTest {
       actualSize++;
     }
     assertEquals(expectedSize, actualSize);
-  }
-
-  private static void removesDateRelatedFields(String dataStoreName, Map<String, Object> document) {
-    if (isMongo(dataStoreName)) {
-      document.remove(MONGO_CREATED_TIME_KEY);
-      document.remove(MONGO_LAST_UPDATED_TIME_KEY);
-      document.remove(MONGO_LAST_UPDATE_TIME_KEY);
-    } else if (isPostgress(dataStoreName)) {
-      document.remove(POSTGRES_CREATED_AT);
-      document.remove(POSTGRES_UPDATED_AT);
-    }
   }
 }
